new manthana radiology/manthana-backend/services/01_body_xray/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.*
python-multipart==0.0.*
torch>=2.0
torchvision>=0.15
timm>=1.0.0
transformers>=4.40
# Pin major line; length assert in txrv_utils catches output-dim drift on upgrade.
torchxrayvision>=1.2
ultralytics>=8.2
numpy>=1.24
Pillow>=10.0
pillow-heif>=0.13.0
opencv-python-headless>=4.8
pydicom>=2.4
# Kimi / Groq narrative (OpenAI-compatible chat.completions API)
openai>=1.40.0
